renamed to InventoryEditUI.java & added the main function as the driver
