--- src/com/neuSep17/ui/InventoryManagementDriver.java
@@ -1,9 +1,0 @@
-package com.neuSep17.ui;
-
-public class InventoryManagementDriver extends InventoryManagementFrame {
-
-  public static void main(String[] args) {
-    InventoryManagementFrame imf=new InventoryManagementFrame();
-  }
-
-}
